+++ aliding-20230426/alibabacloud_aliding20230426/models/__init__.py
@@ -1805,6 +1805,7 @@ from ._create_run_response_body import CreateRunResponseBodyMessagesContentStruc
 from ._create_run_response_body import CreateRunResponseBodyMessagesContentStruct
 from ._create_run_response_body import CreateRunResponseBodyMessages
 from ._create_run_response_body import CreateRunResponseBodyRun
+from ._create_run_response_body import CreateRunResponseBodyThread
 from ._create_scenegroup_headers import CreateScenegroupHeadersAccountContext
 from ._create_schedule_conference_headers import CreateScheduleConferenceHeadersAccountContext
 from ._create_schedule_conference_request import CreateScheduleConferenceRequestScheduleConfSettingModelMoziConfOpenRecordSetting
@@ -4442,6 +4443,7 @@ __all__ = [
     CreateRunResponseBodyMessagesContentStruct,
     CreateRunResponseBodyMessages,
     CreateRunResponseBodyRun,
+    CreateRunResponseBodyThread,
     CreateScenegroupHeadersAccountContext,
     CreateScheduleConferenceHeadersAccountContext,
     CreateScheduleConferenceRequestScheduleConfSettingModelMoziConfOpenRecordSetting,